Fixed some bugs, addressed logical gaps, and corrected wrong imports.

+++ src/main/java/com/example/vibe_store/dto/sale/SaleResponseDTO.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @Setter
 public class SaleResponseDTO {
 
-    private Integer saleId;
+    private Long saleId;
     private Integer employeeId;
     private Integer storeId;
     private BigDecimal saleAmount;
